refactor: resolve error message in ApiMutationError
